[Feat] 거래, 채팅 추가 연동

+++ app/src/main/java/com/example/onenthapp/data/transaction/CompleteTransactionRequest.kt
@@ -1,9 +1,10 @@
 package com.example.onenthapp.data.transaction
 
 data class CompleteTransactionRequest(
-    val dealConfirmationId: Int,
+    val dealConfirmationId: Long,
     val dealDate: String,
     val tradePrice: Int,
     val tradeCount: Int,
     val tradeType: String
 )
+
